requirements.txt
pytest == 7.4.2
selene == 2.0.0rc4
selenium == 4.5.0
allure-pytest==2.13.2
python-dotenv== 1.0.0